Move CUDA/HIP files to a common directory
