+++ src/acc/cuda/PACKAGE
@@ -1,5 +1,5 @@
 {
 "description": "Cuda backend for accelerator API",
 "archive":"libdbcsr",
-"requires": ["../../base", ".."]
+"requires": ["../../base"]
 }
