+++ src/acc/cuda_hip/acc_error.h
@@ -11,7 +11,7 @@
 #define LIBSMM_ACC_H
 
 #if defined(__CUDA)
-# include "acc_cuda.h"
+# include "../cuda/acc_cuda.h"
 #elif defined(__HIP)
 # include "../hip/acc_hip.h"
 #endif
